third_party/blink/web_tests/platform/generic/fast/events/touch/touch-action-touch-handlers-expected.txt

This test verifies the touch event handlers tracked for the compositor for elements with various touch-action settings.

PASS successfullyParsed is true

TEST COMPLETE

Should start with no handlers
PASS getTouchHandlerCount(document) is 0
touch-action: auto should not add any handlers
PASS getTouchHandlerCount(document) is 0
transition from auto should add a handler
PASS getTouchHandlerCount(document) is 1
transition between non-auto values should maintain handler
PASS getTouchHandlerCount(document) is 1
multiple touch-action applications shouldn't affect handler count
PASS getTouchHandlerCount(document) is 1
modifying any unrelated CSS property shouldn't affect handler count
PASS getTouchHandlerCount(document) is 1
setting display:none should remove handler
PASS getTouchHandlerCount(document) is 0
and removing it should bring back handler
PASS getTouchHandlerCount(document) is 1
adding another listener should bump up handler count
PASS getTouchHandlerCount(document) is 2
removing node should remove touch-action handler but not others
PASS getTouchHandlerCount(document) is 1
re-attaching node should add handler
PASS getTouchHandlerCount(document) is 2
transitioning to auto should decrease handler count
PASS getTouchHandlerCount(document) is 1
touch-action on div inside frame should add a handler
PASS getTouchHandlerCount(nestedDocument) is 2
PASS getTouchHandlerCount(document) is 2
